requirements.txt
python-telegram-bot==22.3
requests==2.32.5
pydantic==2.11.7
pydantic-settings==2.10.1

httpx==0.27.0
PyYAML==6.0.2

